268: consolidate model logic into components
