+++ app/adapters/zma.js
@@ -1,17 +1,17 @@
 import DS from 'ember-data';
 import { buildSqlUrl } from '../utils/carto';
 
-const SQL = function(ulurpno) {
-  return `SELECT the_geom, ulurpno as id, project_na, effective, status, lucats FROM zoning_map_amendments WHERE ulurpno='${ulurpno}'`;
+const SQL = function(id) {
+  return `SELECT the_geom, ulurpno as id, project_na, effective, status, lucats FROM zoning_map_amendments WHERE id='${id}'`;
 };
 
 export default DS.JSONAPIAdapter.extend({
   keyForAttribute(key) {
     return key;
   },
-  urlForFindRecord(ulurpno) {
+  urlForFindRecord(id) {
     return buildSqlUrl(
-      SQL(ulurpno),
+      SQL(id),
       'geojson',
     );
   },
