+++ app/components/layer-group-display/special-purpose-district.js
@@ -1,7 +1,24 @@
 import Component from '@ember/component';
-import Bookmarkable from 'labs-zola/mixins/bookmarkable';
+import { computed } from '@ember/object';
+import specialPurposeCrosswalk from 'labs-zola/utils/special-purpose-crosswalk';
+import bbox from '@turf/bbox';
+import { inject as service } from '@ember/service';
 
-const BookmarkableComponent = Component.extend(Bookmarkable);
+export default class LayerGroupDisplaySpecialPurposeDistrictComponent extends Component {
+  @service
+  mainMap;
 
-export default class LayerGroupDisplaySpecialPurposeDistrictComponent extends BookmarkableComponent {
+  @computed('model.value.sdname')
+  get readMoreLink() {
+    const name = this.get('model.value.sdname');
+    const [, [anchorName, boroName]] = specialPurposeCrosswalk
+      .find(([dist]) => dist === name) || [[], []];
+    return `https://www1.nyc.gov/site/planning/zoning/districts-tools/special-purpose-districts-${boroName}.page#${anchorName}`;
+  }
+
+  @computed('model.value.geometry')
+  get bounds() {
+    const geometry = this.get('model.value.geometry');
+    return bbox(geometry);
+  }
 }
