+++ app/models/commercial-overlay.js
@@ -1,21 +1,11 @@
 import DS from 'ember-data';
-import { computed } from '@ember/object';
-import bbox from '@turf/bbox';
-import Bookmarkable from './bookmark';
 
 const { attr } = DS;
 
-export default class MyComponent extends Bookmarkable {
+export default class CommercialOverlay extends DS.Model {
   @attr()
   geometry;
 
   @attr('string')
   overlay;
-
-  @computed('geometry')
-  get bounds() {
-    const geometry = this.get('geometry');
-
-    return bbox(geometry);
-  }
 }
