+++ app/templates/components/grouped-checkboxes.hbs
@@ -20,7 +20,7 @@
   {{/if}}
 
   <ul class="nested columns-3 hidable-sibling" data-test-grouped={{group.name}}>
-    {{#each group.codes as |code|}} 
+    {{#each group.codes as |code|}}
       <li>
         <label>
           <input type='checkbox'
