+++ app/templates/components/layer-group-display/special-purpose-subdistrict.hbs
@@ -1,16 +1,16 @@
 <label class="header-type">
-  Special Purpose District <span class="medium-gray">|</span> {{model.value.sdlbl}}
+  Special Purpose Subdistrict <span class="medium-gray">|</span> {{model.value.splbl}}
 </label>
 
 <h1 class="header-large">
-  {{model.value.sdname}}
+  {{model.value.spname}}
 </h1>
 
 <p>Since 1969, the City Planning Commission has designated special zoning districts in response to areas of the City with unique characteristics. Each special district stipulates zoning requirements and incentives tailored to specific conditions that may not lend themselves to generalized zoning and standard development.</p>
 
   <p>
     <a
-      href="{{readMoreLink}}"
+      href="https://www1.nyc.gov/site/planning/zoning/districts-tools/special-purpose-districts-citywide.page"
       target="_blank"
     >
       {{fa-icon 'external-link-alt'}} Learn more&hellip;</a></p>
