+++ app/templates/components/layer-group-display/zoning-map-amendment.hbs
@@ -1,30 +1,23 @@
-<div class="content-close-button-container">{{#link-to 'index' classNames='close-button'}}<span aria-hidden="true">&times;</span>{{/link-to}}</div>
+<label class="header-type">
+  ZONING MAP AMENDMENT
+</label>
 
-<div class="content-area cell large-5 large-cell-block-y xxlarge-4">
-  {{#if model.isRunning}}
-    {{#content-placeholders as |placeholder|}}
-      {{placeholder.text lines=1}}
-      {{placeholder.nav}}
-      {{placeholder.text lines=10}}
-    {{/content-placeholders}}
-  {{else}}
-    {{bookmark-button
-      bookmark=model.value.bookmark
-      createBookmark=(action 'createBookmark')}}
+{{log model.value.project_na}}
 
-    <label class="header-label">ZONING MAP AMENDMENT</label>
-    <h1 class="header-large">{{model.value.project_na}}</h1>
+<h1 class="header-large">
+  {{model.value.project_na}}
+</h1>
 
-    <div class="data-grid"><label class="data-label">ULURP Number</label><span class="datum"><a target="_blank" href="https://zap-api.planninglabs.nyc/projects/ulurp/{{model.value.lucats}}">{{fa-icon 'external-link-alt'}} {{model.value.id}}</a></span></div>
-    <div class="data-grid"><label class="data-label">Effective</label>
-      <span class="datum">
-        {{#async-await model.value.effectiveDisplay as |date|}} 
-          {{date}}
-        {{/async-await}}
-      </span>
-    </div>
-    <div class="data-grid"><label class="data-label">Status</label><span class="datum">{{model.value.status}}</span></div>
-    <div class="data-grid"><label class="data-label">Zoning Sketch Map</label><span class="datum"><a target="_blank" href="http://www1.nyc.gov/assets/planning/download/pdf/zoning/zoning-maps/sketchmaps/skz{{model.value.id}}.pdf">{{fa-icon "file-pdf-o"}} {{model.value.id}} <small>(PDF)</small></a></span></div>
-  {{/if}}
-
+<div class="data-grid"><label class="data-label">ULURP Number</label><span class="datum"><a target="_blank" href="https://zap-api.planninglabs.nyc/projects/ulurp/{{model.value.lucats}}">{{fa-icon 'external-link-alt'}} {{model.value.id}}</a></span></div>
+<div class="data-grid"><label class="data-label">Effective</label>
+  <span class="datum">
+    {{#async-await model.value.effectiveDisplay as |date|}}
+      {{date}}
+    {{/async-await}}
+  </span>
 </div>
+
+<div class="data-grid"><label class="data-label">Status</label><span class="datum">{{model.value.status}}</span></div>
+<div class="data-grid"><label class="data-label">Zoning Sketch Map</label><span class="datum"><a target="_blank" href="http://www1.nyc.gov/assets/planning/download/pdf/zoning/zoning-maps/sketchmaps/skz{{model.value.project_na}}.pdf">{{fa-icon "file-pdf-o"}} {{model.value.project_na}} <small>(PDF)</small></a></span></div>
+
+{{outlet}}
